fix: handle a race condition
Depending on the time it takes to complete the async code, it may be
that a new async operation is started, which finishes before the
original one. In that case, the original one (the one started first)
will overwrite the one started second.

This change fixes this by applying a version number, ensuring that
actions which where started first, never overwrite results which were
started later.

This is accomplished by a simple 64bit version number, which is
incremented consecutively. In theory that may wrap around at some point.
Still, in practice, it is highly unlikely.